Fix : DetailContainer를 Addtional,Comments,Detail로 세분화 (Redux store상태값 배분을 위함)
