+++ server/view/index.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="minimum-scale=1, width=device-width, initial-scale=1.0">
-    <title> NEW SMART </title>
+    <title> Where | Tour Guide </title>
 </head>
 <body>
     <div id='root'></div>
-<script src="react.bundle.5bc1d3b3191d3e6b6d23.js"></script><script src="vendors.5bc1d3b3191d3e6b6d23.js"></script><script src="app.5bc1d3b3191d3e6b6d23.js"></script></body>
+<script src="react.bundle.cc2d72d372e71d548663.js"></script><script src="vendors.cc2d72d372e71d548663.js"></script><script src="app.cc2d72d372e71d548663.js"></script></body>
 </html>
