+++ src/detail/component/Detail.js
@@ -1,27 +1,19 @@
 import styled from "styled-components";
-import { Avatar, Paper, IconButton, Divider, Chip } from "@material-ui/core";
+import { Avatar, Divider } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import { Favorite as FavoriteIcon, Share as ShareIcon } from '@material-ui/icons';
 import { blue, red } from "@material-ui/core/colors";
-import { withRouter } from "react-router-dom";
-import isInProgress from '../../common/isInProgressDate';
-import { getCategory } from '../../common/categoryCode';
-import { callApiDetailIntro } from '../../common/api';
-import Overview from './Overview';
-import AdditionalInfo from './Additional';
-import CommentsInput from './CommentsInput';
-import Comments from './Comments';
-import Map from './Map';
 
-const DetailContainer = styled(Paper)`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  min-width: 1300px;
-  position: relative;
-  padding: 30px;
-`;
+
+
+// const DetailContainer = styled(Paper)`
+//   display: flex;
+//   flex-direction: column;
+//   justify-content: center;
+//   align-items: center;
+//   min-width: 1300px;
+//   position: relative;
+//   padding: 30px;
+// `;
 const HeaderContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -99,14 +91,7 @@ const Badge = styled.span`
   margin-top: 14px;
   margin-right: 5px;
 `;
-const ActionsWrap = styled.div`
-  width: 100%;
-  margin-top: 50px;
-  margin-right: 50px;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-`;
+
 const ShortDivider = styled(Divider)`
   border: 2px;
   width: 1000px;
@@ -140,14 +125,7 @@ const InfoWrap = styled.div`
   font-size: 1.2rem;
   font-weight: bold;
 `;
-const MapContainer = styled.div`
-  padding: 50px 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 600px;
-`;
+
 const useStyles = makeStyles((theme) => ({
   green: {
     position: "absolute",
@@ -177,51 +155,22 @@ const useStyles = makeStyles((theme) => ({
 
 
 
-function Detail( props ) {
+export default function Detail( props ) {
   const classes = useStyles();
-  const {place, handleScrap, handleShare, comments, match} = props;
+  const {place, additionalInfo, category} = props;
   const {firstimage, title, date, addr1, tel, dist, readcount} = place;
-  const {contentid, contenttypeid} = match.params;
-
-  const [additionalInfo, setAdditionalInfo] = React.useState({
-    origin: {},
-    destination: {},
-    isInProgress: false,
-    additional: []
-  });
-
-  React.useEffect(() => {
-    callApiDetailIntro(contenttypeid, contentid)
-      .then(res => {
-        console.log(`DetailCommonIntro API 호출 ${contenttypeid} ${contentid}`);
-
-        setAdditionalInfo({
-          origin: JSON.parse(sessionStorage.getItem("location")), 
-          destination: {
-            lat: res.data.mapy,
-            lng: res.data.mapx
-          },
-          overview: res.data.overview,
-          inProgress: isInProgress(res.eventstartdate, res.eventenddate),
-          additional: Object.entries(res.data)
-        });
-      })
-      .catch(err => {
-        console.log(err);
-      })
-  }, []);
 
   return (
-    <DetailContainer elevation={3}>
+    <React.Fragment>
       <HeaderContainer>
         <TitleContainer>
           <CategoryWrap>
-            {getCategory(contenttypeid)}
+            {category}
           </CategoryWrap>
           <TitleWrap> {title} </TitleWrap>
           <BadgeWrap>
             <DateWrap> {date} </DateWrap>
-            {additionalInfo.isInProgress && <Badge color="blue"> 진행중 </Badge> }
+            {/* {additionalInfo.isInProgress && <Badge color="blue"> 진행중 </Badge> } */}
             {readcount >=2000 && <Badge color="red"> 인기 </Badge> }
             {dist < 1000 && <Badge color="green"> 가까움 </Badge> }
           </BadgeWrap>
@@ -239,30 +188,7 @@ function Detail( props ) {
         </Avatar>
         <Image src={firstimage} />
       </ImageContainer>
-
-      <Overview description={additionalInfo.overview} />
-
-      <AdditionalInfo additional={additionalInfo.additional} />
-
-      <MapContainer>
-        <Map 
-        origin={additionalInfo.origin} 
-        destination={additionalInfo.destination} />
-      </MapContainer>
-
-      <ActionsWrap>
-          <IconButton aria-label="add to favorites" onClick={handleScrap} style={{fontSize: "1.6rem"}}>
-              <FavoriteIcon/>
-          </IconButton>
-          <IconButton aria-label="share" onClick={handleShare} style={{fontSize: "1.6rem"}}>
-              <ShareIcon />
-          </IconButton>
-      </ActionsWrap>
-      
-      <CommentsInput />
-      <Comments comments={comments}/>
-    </DetailContainer>
+    </React.Fragment>
   );
 };
 
-export default withRouter(Detail);
