+++ src/detail/component/Map.js
@@ -51,11 +51,6 @@ function toFloatLatLng(obj) {
     const lng  = parseFloat(obj.lng);
     return { lat, lng };
 }
-function toCoordinatesString(obj) {
-    const coordinates = Object.values(obj).join(',');
-    console.log(coordinates)
-    return coordinates;
-}
 
 export default function Map(props) {
     const {origin, destination} = props;
