+++ src/detail/page/DetailPage.js
@@ -1,12 +1,30 @@
-import DetailContainer from "../container/DetailContainer";
 import styled from "styled-components";
+import { Paper } from '@material-ui/core';
+import DetailContainer from "../container/DetailContainer";
+import AdditionalContainer from '../container/AdditionalContainer';
+import CommentsContainer from '../container/CommentsContainer';
 
-const DetailPageContainer = styled.section``;
+const Container = styled(Paper)`
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  min-width: 1300px;
+  position: relative;
+  padding: 30px;
+`;
+
+
 
 export default function DetailPage() {
+
+  const [place, setPlace] = React.useState(JSON.parse(sessionStorage.getItem("currentPlace")));
+
   return (
-    <DetailPageContainer>
-      <DetailContainer />
-    </DetailPageContainer>
+    <Container>
+      <DetailContainer place={place}/>
+      <AdditionalContainer />
+      <CommentsContainer />
+    </Container>
   );
 }
